[refactor] 그룹 관련 패키지 분리, 바인딩 어댑터 통합
- 그룹 선택과 그룹 추가 각각을 별도의 패키지로 구분하고 사용하던 바인딩
어댑터를 util 패키지의 BindingAdapters로 통합함.
