+++ mobile/app/src/main/java/com/lacuc/pets/MainModule.kt
@@ -1,7 +1,7 @@
 package com.lacuc.pets
 
-import com.lacuc.pets.ui.group.AddGroupFragment
-import com.lacuc.pets.ui.group.ChooseGroupFragment
+import com.lacuc.pets.ui.group.add.AddGroupFragment
+import com.lacuc.pets.ui.group.choose.ChooseGroupFragment
 import com.lacuc.pets.ui.login.signup.SignUpFragment
 import dagger.Module
 import dagger.android.ContributesAndroidInjector
