+++ mobile/app/src/main/java/com/lacuc/pets/di/ViewModelModule.kt
@@ -3,8 +3,8 @@ package com.lacuc.pets.di
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import com.lacuc.pets.ViewModelFactory
-import com.lacuc.pets.ui.group.AddGroupViewModel
-import com.lacuc.pets.ui.group.ChooseGroupViewModel
+import com.lacuc.pets.ui.group.add.AddGroupViewModel
+import com.lacuc.pets.ui.group.choose.ChooseGroupViewModel
 import com.lacuc.pets.ui.login.signup.SignUpViewModel
 import dagger.Binds
 import dagger.MapKey
