--- mobile/app/src/main/java/com/lacuc/pets/ui/group/GroupItemBinding.kt
@@ -1,12 +1,0 @@
-package com.lacuc.pets.ui.group
-
-import android.widget.ImageView
-import androidx.databinding.BindingAdapter
-import com.bumptech.glide.Glide
-
-@BindingAdapter("app:image")
-fun setImage(imageView: ImageView, url: String) {
-    Glide.with(imageView.context)
-        .load(url)
-        .into(imageView)
-}
