+++ mobile/app/src/main/java/com/lacuc/pets/util/BindingAdpaters.kt
@@ -1,10 +1,20 @@
-package com.lacuc.pets.ui.group
+package com.lacuc.pets.util
 
 import android.view.ViewGroup
+import android.widget.ImageView
 import androidx.databinding.BindingAdapter
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
+import com.bumptech.glide.Glide
 import com.lacuc.pets.domain.group.GroupItem
+import com.lacuc.pets.ui.group.choose.ChooseGroupAdapter
+
+@BindingAdapter("app:image")
+fun setImage(imageView: ImageView, url: String) {
+    Glide.with(imageView.context)
+        .load(url)
+        .into(imageView)
+}
 
 @BindingAdapter("app:items")
 fun setItems(listView: RecyclerView, items: List<GroupItem>) {
