+++ mobile/app/src/main/res/layout/fragment_choose_group.xml
@@ -2,7 +2,7 @@
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".ui.group.ChooseGroupFragment">
+    tools:context=".ui.group.choose.ChooseGroupFragment">
 
     <data>
 
@@ -10,7 +10,7 @@
 
         <variable
             name="vm"
-            type="com.lacuc.pets.ui.group.ChooseGroupViewModel" />
+            type="com.lacuc.pets.ui.group.choose.ChooseGroupViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
